Release 1.20-1.1.1 -Reworked Tier Upgrades, fixed all bugs related to missing items -Added an Auto-Output button for Cobblestone Generator GUI -Now, all faces of furnace-like blocks with a Liquid Fuel Upgrade will be set to this upgrade tank, if a Generator Upgrade or Experience Upgrade isn't used -Updated Factory API to 1.16.5-2.0

+++ common/src/main/java/wily/betterfurnaces/client/screen/ColorUpgradeScreen.java
@@ -65,7 +65,7 @@ public class ColorUpgradeScreen extends AbstractUpgradeScreen<ContainerColorUpgr
         ItemColorsHandler.putColor(tag,red.getValue(),green.getValue(),blue.getValue());
         stack.setTag(tag);
         graphics.pose().pushPose();
-        graphics.pose().translate((width / 2 ), (this.topPos - 40),0);
+        graphics.pose().translate(((float) width / 2 ), (this.topPos - 40),0);
         graphics.pose().scale(4F,4F,1F);
         graphics.pose().translate(-8, -8,0);
         graphics.renderItem(stack, 0, 0);
